refactor: move cli to cli and entrypoint to main.
